added few comments and fixed temp in history
