+++ src/Components/Nav.css
@@ -1,5 +1,4 @@
-@import
-url(https://fonts.googleapis.com/css?family-Montserrat);
+
 
 .navbar {
   height: 100px;
